+++ src/main/java/duke/Duke.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.command.Command;
+
 public class Duke {
     private Storage storage;
     private TaskList tasks;
