+++ src/main/java/duke/Parser.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.command.*;
+
 public class Parser {
     public static Command parse(String str) {
         Command command;
@@ -17,28 +21,28 @@ public class Parser {
             if (deadlineCheck) {
                 String target = " /by ";
                 if (!str.contains(target)) {
-                    throw new RuntimeException("Unable to create Deadline! Deadline commands need a /by field!");
+                    throw new RuntimeException("Unable to create duke.task.Deadline! duke.task.Deadline commands need a /by field!");
                 }
                 int index = str.indexOf(target);
                 String description = str.substring(0, index);
                 if (description.equals("")) {
-                    throw new RuntimeException("Unable to create Deadline! Description for deadline cannot be left blank!");
+                    throw new RuntimeException("Unable to create duke.task.Deadline! Description for deadline cannot be left blank!");
                 }
                 String deadline = str.substring(index + 5);
                 int dateTimeLength = deadline.length();
                 if (!(dateTimeLength > 12 && dateTimeLength < 16)) {
-                    throw new RuntimeException("Unable to create Deadline! Check your date and time. They have to be in the format of dd/mm/yyyy hhmm");
+                    throw new RuntimeException("Unable to create duke.task.Deadline! Check your date and time. They have to be in the format of dd/mm/yyyy hhmm");
                 }
                 int firstSlash = deadline.indexOf("/");
                 int secondSlash = deadline.indexOf("/", firstSlash + 1);
                 if (firstSlash == -1 || secondSlash == -1) {
-                    throw new RuntimeException("Unable to create Deadline! Check your date format. Use / to separate day, month and year.");
+                    throw new RuntimeException("Unable to create duke.task.Deadline! Check your date format. Use / to separate day, month and year.");
                 }
             } else if (eventCheck) {
                 String target1 = " /from ";
                 String target2 = " /to ";
                 if (!(str.contains(target1) && str.contains(target2))) {
-                    throw new RuntimeException("Unable to create Event! Event commands need a /from and /to field!");
+                    throw new RuntimeException("Unable to create duke.task.Event! duke.task.Event commands need a /from and /to field!");
                 }
                 int index1 = str.indexOf(target1);
                 int index2 = str.indexOf(target2);
@@ -58,7 +62,7 @@ public class Parser {
                 int endFirstSlash = end.indexOf("/");
                 int endSecondSlash = end.indexOf("/", endFirstSlash + 1);
                 if (startFirstSlash == -1 || startSecondSlash == -1 || endFirstSlash == -1 || endSecondSlash == -1) {
-                    throw new RuntimeException("Unable to create Deadline! Check your date and time format!");
+                    throw new RuntimeException("Unable to create duke.task.Deadline! Check your date and time format!");
                 }
             }
             command = new AddCommand(str);
