+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.Task;
+
 public class TaskList {
     private int index;
     private Task[] tasks;
@@ -31,7 +35,7 @@ public class TaskList {
 
     public String deleteTask(int index) {
         if (index < 0 || index > 99 || tasks[index] == null) {
-            throw new RuntimeException("Task does not exist!");
+            throw new RuntimeException("duke.task.Task does not exist!");
         }
         String str = "Got it! This task has been removed:\n";
         str += this.printTask(index) + "\n";
@@ -47,7 +51,7 @@ public class TaskList {
 
     public String markTask(int index) {
         if (index < 0 || index > 99 || tasks[index] == null) {
-            throw new RuntimeException("Task does not exist!");
+            throw new RuntimeException("duke.task.Task does not exist!");
         }
         this.tasks[index].mark();
         String str = "Great job! This task has been marked as done:\n";
@@ -57,7 +61,7 @@ public class TaskList {
 
     public String unMarkTask(int index) {
         if (index < 0 || index > 99 || tasks[index] == null) {
-            throw new RuntimeException("Task does not exist!");
+            throw new RuntimeException("duke.task.Task does not exist!");
         }
         this.tasks[index].unMark();
         String str = "Noted! This task has been marked as undone:\n";
