+++ src/main/java/duke/TaskType.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum TaskType {
     TODO, DEADLINE, EVENT;
