+++ src/main/java/duke/command/UnMarkCommand.java
@@ -1,13 +1,20 @@
-public class DeleteCommand extends Command {
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class UnMarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public UnMarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        System.out.println(taskList.deleteTask(this.index));
+        System.out.println(taskList.unMarkTask(this.index));
         storage.save(taskList);
     }
