+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.TaskType;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
